Full code refactoring, separated everything in modules updated readme, ready to merge for releasing superduper alpha
