+++ AniMate/detection/__init__.py
@@ -1,0 +1,8 @@
+"""
+AniMate detection package: provides MediaPipe-based pose, hands, and face tracking.
+"""
+from .mediapipe_detector import MediaPipeDetector
+
+__all__ = [
+    "MediaPipeDetector",
+]
