+++ data/__init__.py
@@ -12,11 +12,6 @@ from .landmark_structure import (
     LANDMARK_CONNECTIONS
 )
 
-from .test_rigs import (
-    BASIC_HUMANOID,
-    FACE_RIG,
-    HAND_RIG
-)
 
 from .bone_mappings import (
     POSE_LANDMARKS_TO_BONES,
@@ -29,9 +24,6 @@ __all__ = [
     'FACE_LANDMARKS',
     'HAND_LANDMARKS',
     'LANDMARK_CONNECTIONS',
-    'BASIC_HUMANOID',
-    'FACE_RIG',
-    'HAND_RIG',
     'POSE_LANDMARKS_TO_BONES',
     'FACE_LANDMARKS_TO_BONES',
     'HAND_LANDMARKS_TO_BONES'
