--- data/mixamorig_bone_hierarchy.json
@@ -1,99 +1,0 @@
-{
-    "Hips": {
-        "Spine": {
-            "Spine1": {
-                "Spine2": {
-                    "Neck": {
-                        "Head": {
-                            "HeadTop_End": {}
-                        }
-                    }
-                }
-            }
-        },
-        "LeftUpLeg": {
-            "LeftLeg": {
-                "LeftFoot": {
-                    "LeftToeBase": {
-                        "LeftToe_End": {}
-                    }
-                }
-            }
-        },
-        "RightUpLeg": {
-            "RightLeg": {
-                "RightFoot": {
-                    "RightToeBase": {
-                        "RightToe_End": {}
-                    }
-                }
-            }
-        },
-        "LeftShoulder": {
-            "LeftArm": {
-                "LeftForeArm": {
-                    "LeftHand": {
-                        "LeftHandThumb1": {
-                            "LeftHandThumb2": {
-                                "LeftHandThumb3": {}
-                            }
-                        },
-                        "LeftHandIndex1": {
-                            "LeftHandIndex2": {
-                                "LeftHandIndex3": {}
-                            }
-                        },
-                        "LeftHandMiddle1": {
-                            "LeftHandMiddle2": {
-                                "LeftHandMiddle3": {}
-                            }
-                        },
-                        "LeftHandRing1": {
-                            "LeftHandRing2": {
-                                "LeftHandRing3": {}
-                            }
-                        },
-                        "LeftHandPinky1": {
-                            "LeftHandPinky2": {
-                                "LeftHandPinky3": {}
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        "RightShoulder": {
-            "RightArm": {
-                "RightForeArm": {
-                    "RightHand": {
-                        "RightHandThumb1": {
-                            "RightHandThumb2": {
-                                "RightHandThumb3": {}
-                            }
-                        },
-                        "RightHandIndex1": {
-                            "RightHandIndex2": {
-                                "RightHandIndex3": {}
-                            }
-                        },
-                        "RightHandMiddle1": {
-                            "RightHandMiddle2": {
-                                "RightHandMiddle3": {}
-                            }
-                        },
-                        "RightHandRing1": {
-                            "RightHandRing2": {
-                                "RightHandRing3": {}
-                            }
-                        },
-                        "RightHandPinky1": {
-                            "RightHandPinky2": {
-                                "RightHandPinky3": {}
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
