--- rig/__init__.py
@@ -1,5 +1,0 @@
-__author__ = "Valentin Bakin"
-
-"""
-Rig package for handling the 3D rig and animation logic.
-"""
