--- rig/mappings/maya_mapping.py
@@ -1,42 +1,0 @@
-"""
-Maya rig mapping implementation.
-"""
-from .base_mappings import BaseRigMapping
-
-class MayaMapping(BaseRigMapping):
-    """Mapping for Maya rigs (placeholder)."""
-
-    def get_bone_hierarchy(self):
-        """Get the expected bone hierarchy for Maya rigs."""
-        return {}  # Empty for now
-
-    def get_pose_mapping(self):
-        """Get the mapping between MediaPipe pose landmarks and bone names."""
-        return {}  # Empty for now
-
-    def get_face_mapping(self):
-        """Get the mapping between MediaPipe face landmarks and bone names."""
-        return {}  # Empty for now
-
-    def get_hand_mapping(self):
-        """Get the mapping between MediaPipe hand landmarks and bone names."""
-        return {}  # Empty for now
-
-    def get_bone_rotation_limits(self):
-        """Get rotation limits for each bone."""
-        return {}  # Empty for now
-
-    def get_bone_scale_factors(self):
-        """Get scale factors for each bone's rotation."""
-        return {}  # Empty for now
-
-    def get_capabilities(self):
-        """Get the capabilities of this rig type."""
-        return {
-            'face': False,
-            'hands': False,
-        } 
-
-    def get_axis_corrections(self):
-        return {
-        }
